--- library/item/item_rendering/build.gradle
@@ -1,20 +1,0 @@
-plugins {
-	id("qsl.module")
-}
-
-qslModule {
-	name = "Quilt Item Rendering API"
-	moduleName = "item_rendering"
-	id = "quilt_item_rendering"
-	description = "An API for modifying and extending the rendering of items in inventories."
-	library = "item"
-	moduleDependencies {
-		core {
-			api("qsl_base")
-			testmodOnly("resource_loader")
-		}
-	}
-	injectedInterface("net/minecraft/class_1792") {
-		values = ["org/quiltmc/qsl/item/rendering/api/client/QuiltItemRenderingExtensions"]
-	}
-}
