--- library/item/item_rendering/src/main/resources/quilt_item_rendering.mixins.json
@@ -1,11 +1,0 @@
-{
-  "required": true,
-  "package": "org.quiltmc.qsl.item.rendering.mixin",
-  "compatibilityLevel": "JAVA_17",
-  "client": [
-    "client.ItemRendererMixin"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
