--- library/item/item_rendering/src/testmod/java/org/quiltmc/qsl/item/rendering/test/client/itembar/ConstantItemBarRenderer.java
@@ -1,40 +1,0 @@
-/*
- * Copyright 2022 QuiltMC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.quiltmc.qsl.item.rendering.test.client.itembar;
-
-import net.minecraft.item.ItemStack;
-
-import org.quiltmc.qsl.item.rendering.api.client.SolidColorItemBarRenderer;
-
-public class ConstantItemBarRenderer extends SolidColorItemBarRenderer {
-	protected final int step, color;
-
-	public ConstantItemBarRenderer(int step, int color) {
-		this.step = step;
-		this.color = color;
-	}
-
-	@Override
-	protected int getItemBarStep(ItemStack stack) {
-		return step;
-	}
-
-	@Override
-	protected int getItemBarForeground(ItemStack stack) {
-		return color;
-	}
-}
