+++ library/rendering/build.gradle
@@ -1,0 +1,7 @@
+plugins {
+	id("qsl.library")
+}
+
+qslLibrary {
+	libraryName = "rendering"
+}
