+++ library/rendering/item_rendering/build.gradle
@@ -1,0 +1,20 @@
+plugins {
+	id("qsl.module")
+}
+
+qslModule {
+	name = "Quilt Item Rendering API"
+	moduleName = "item_rendering"
+	id = "quilt_item_rendering"
+	description = "An API for modifying and extending the rendering of items in inventories."
+	library = "rendering"
+	moduleDependencies {
+		core {
+			api("qsl_base")
+			testmodOnly("resource_loader")
+		}
+	}
+	injectedInterface("net/minecraft/class_1792") {
+		values = ["org/quiltmc/qsl/rendering/item/api/client/QuiltItemRenderingExtensions"]
+	}
+}
