+++ library/rendering/item_rendering/src/main/resources/assets/quilt_item_rendering/icon.png
