+++ library/rendering/item_rendering/src/main/resources/quilt_item_rendering.mixins.json
@@ -1,0 +1,11 @@
+{
+  "required": true,
+  "package": "org.quiltmc.qsl.rendering.item.mixin",
+  "compatibilityLevel": "JAVA_17",
+  "client": [
+    "client.ItemRendererMixin"
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
