+++ library/rendering/item_rendering/src/testmod/java/org/quiltmc/qsl/rendering/item/test/CooldownItem.java
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2022 QuiltMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.quiltmc.qsl.rendering.item.test;
+
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.item.Item;
+import net.minecraft.item.ItemStack;
+import net.minecraft.util.Hand;
+import net.minecraft.util.TypedActionResult;
+import net.minecraft.world.World;
+
+public class CooldownItem extends Item {
+	// Cooldown duration in ticks when this item is being used
+	private final int duration;
+
+	public CooldownItem(Settings settings, int duration) {
+		super(settings);
+		this.duration = duration;
+	}
+
+	@Override
+	public TypedActionResult<ItemStack> use(World world, PlayerEntity user, Hand hand) {
+		user.getItemCooldownManager().set(this, duration);
+		return TypedActionResult.success(user.getStackInHand(hand));
+	}
+}
