+++ library/rendering/item_rendering/src/testmod/resources/assets/quilt_item_rendering_testmod/models/item/mana_storage.json
