+++ library/rendering/item_rendering/src/testmod/resources/quilt.mod.json
@@ -1,0 +1,33 @@
+{
+  "schema_version": 1,
+  "quilt_loader": {
+    "group": "org.quiltmc.qsl.item.rendering",
+    "id": "quilt_item_rendering_testmod",
+    "version": "1.0.0",
+    "metadata": {
+      "name": "Quilt Item Rendering API Test Mod",
+      "description": "Testmod for Quilt APIs relating to item rendering.",
+      "license": "Apache-2.0",
+      "contact": {
+        "homepage": "https://quiltmc.org",
+        "issues": "https://github.com/QuiltMC/quilt-standard-libraries/issues",
+        "sources": "https://github.com/QuiltMC/quilt-standard-libraries"
+      },
+      "contributors": {
+        "QuiltMC: QSL Team": "Author"
+      }
+    },
+    "intermediate_mappings": "net.fabricmc:intermediary",
+    "load_type": "always",
+    "entrypoints": {
+      "init": [
+        "org.quiltmc.qsl.rendering.item.test.ItemRenderingTestmod"
+      ]
+    },
+    "depends": [
+      "quilt_loader",
+      "quilt_resource_loader",
+      "quilt_item_rendering"
+    ]
+  }
+}
