+++ library/rendering/src/testmod/resources/quilt.mod.json
@@ -1,0 +1,14 @@
+{
+  "schema_version": 1,
+  "quilt_loader": {
+    "group": "org.quiltmc.qsl",
+    "id": "quilt_rendering_testmod",
+    "version": "1.0.0",
+    "metadata": {
+      "name": "Quilt Rendering Library Test Mod",
+      "license": "Apache-2.0"
+    },
+    "intermediate_mappings": "net.fabricmc:intermediary",
+    "load_type": "always"
+  }
+}
