+++ settings.gradle
@@ -29,6 +29,7 @@ includeBuild("build-logic")
 library("core")
 library("block")
 library("data")
+library("rendering")
 library("gui")
 library("item")
 library("management")
